backend/requirements/base.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
psycopg[binary]>=3.2,<4.0
redis>=6.0,<7.0
celery>=5.5,<6.0
django-celery-results>=2.6,<3.0
requests>=2.32,<3.0
python-dotenv>=1.0,<2.0
pgvector>=0.3.6,<1.0
google-genai>=1.0.0